feat: move utility functions to runtime_utils and update imports accordingly
